+++ CHANGELOG.md
@@ -7,3 +7,7 @@ This significantly reduces the maintenance effort and makes it easier to use fea
 
 EasyArmorStands v2 can still be used on Bukkit/Spigot/Paper 1.8.8 to 1.21.11.
 However, most new features will only be added to EasyArmorStands v3.
+
+# Improvements
+
+* Add Mannequin pose property